# backend/docs/swagger.yaml
basePath: /
definitions:
  dto.CreateProductRequest:
    properties:
      bar_code:
        example: 7501234567890
        type: integer
      category:
        example: Remeras
        type: string
      description:
        example: Remera de algodón 100% color negro, cuello redondo
        type: string
      size:
        description: S,M,L,XL,XXL
        enum:
        - S
        - M
        - L
        - XL
        - XXL
        example: M
        type: string
      stock:
        example: 50
        minimum: 0
        type: integer
      title:
        example: Remera Básica Negra
        type: string
      unit_price:
        example: 2500
        minimum: 0
        type: number
    required:
    - bar_code
    - category
    - description
    - size
    - stock
    - title
    - unit_price
    type: object
  dto.ErrorGeneral:
    properties:
      meesage:
        example: message error
        type: string
    type: object
  dto.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: secret123
        type: string
    type: object
  dto.LoginResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR...
        type: string
      expires_in:
        example: 3600
        type: integer
      token_type:
        example: Bearer
        type: string
    type: object
  dto.ProductImageResponse:
    properties:
      created_at:
        example: "2025-01-15T10:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      is_primary:
        example: true
        type: boolean
      position:
        example: 0
        type: integer
      product_id:
        example: 1
        type: integer
      url:
        example: /static/products/1/image.jpg
        type: string
    type: object
  dto.ProductResponse:
    properties:
      bar_code:
        example: 7501234567890
        type: integer
      category:
        example: Remeras
        type: string
      description:
        example: Remera de algodón 100% color negro, cuello redondo
        type: string
      id:
        example: 1
        type: integer
      size:
        example: M
        type: string
      stock:
        example: 50
        type: integer
      title:
        example: Remera Básica Negra
        type: string
      unit_price:
        example: 2500
        type: number
    type: object
  dto.RegisterRequest:
    properties:
      email:
        example: juan@example.com
        type: string
      first_name:
        example: Juan
        type: string
      last_name:
        example: Pérez
        type: string
      password:
        example: "123456"
        type: string
    type: object
  dto.RegisterResponse:
    properties:
      token:
        description: JWT
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
  dto.UpdateProductRequest:
    properties:
      bar_code:
        example: 7501234567890
        type: integer
      category:
        example: Remeras
        type: string
      description:
        example: Remera de algodón 100% color negro, cuello redondo
        type: string
      size:
        description: S,M,L,XL,XXL
        example: M
        type: string
      stock:
        example: 50
        type: integer
      title:
        example: Remera Básica Negra
        type: string
      unit_price:
        example: 2500
        type: number
    type: object
  dto.UserResponse:
    properties:
      email:
        example: juan@example.com
        type: string
      first_name:
        example: Juan
        type: string
      id:
        example: 1
        type: integer
      last_name:
        example: Pérez
        type: string
      provider:
        example: local
        type: string
      role:
        example: user
        type: string
    type: object
host: localhost:9090
info:
  contact: {}
  description: API for managing products with authentication
  title: Core API
  version: "1.0"
paths:
  /api/auth/google/login:
    get:
      description: Redirecciona al usuario a la página de login de Google.
      produces:
      - application/json
      responses:
        "302":
          description: Redirección a Google
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorGeneral'
      summary: Login con Google (redirect)
      tags:
      - auth
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Autenticación local. Devuelve JWT y datos básicos del usuario.
      parameters:
      - description: Credenciales de acceso
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorGeneral'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorGeneral'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorGeneral'
      summary: Login (email y password)
      tags:
      - auth
  /api/auth/me:
    get:
      description: Devuelve la información del usuario actual basada en el JWT.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorGeneral'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorGeneral'
      security:
      - BearerAuth: []
      summary: Obtener usuario autenticado
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Registra un nuevo usuario con email y contraseña y devuelve JWT
        + datos del usuario.
      parameters:
      - description: Datos de registro
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorGeneral'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorGeneral'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorGeneral'
      summary: Registrar usuario local
      tags:
      - auth
  /api/products:
    get:
      description: Filtra por categoría, talla, query, limit y offset
      parameters:
      - description: Categoría
        in: query
        name: category
        type: string
      - description: Talla (S,M,L,XL,XXL)
        in: query
        name: size
        type: string
      - description: Búsqueda en título/desc
        in: query
        name: q
        type: string
      - description: Límite (<=100)
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ProductResponse'
            type: array
      summary: Listar productos
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Crea un nuevo producto (solo admin)
      parameters:
      - description: Product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorGeneral'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorGeneral'
      security:
      - BearerAuth: []
      summary: Crear producto
      tags:
      - products
  /api/products/{id}:
    delete:
      description: Elimina un producto por ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorGeneral'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorGeneral'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorGeneral'
      security:
      - BearerAuth: []
      summary: Eliminar producto
      tags:
      - products
    get:
      description: Obtiene un producto específico por su ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obtener producto por ID
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Actualiza un producto existente por ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Actualizar producto
      tags:
      - products
  /api/products/{id}/images:
    get:
      description: Get all images associated with a product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ProductImageResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all images for a product
      tags:
      - products
    post:
      consumes:
      - multipart/form-data
      description: Upload an image file for a specific product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Image file
        in: formData
        name: file
        required: true
        type: file
      - description: Is primary image
        in: formData
        name: is_primary
        type: boolean
      - description: Image position
        in: formData
        name: position
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ProductImageResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload an image for a product
      tags:
      - products
  /api/products/{id}/images/{imageId}:
    delete:
      description: Delete a specific product image by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Image ID
        in: path
        name: imageId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a product image
      tags:
      - products
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"